requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
celery
redis
python-dotenv
werkzeug